+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SpiceKernels can now load SPK files directly, without needing the files to be in the
   cache.
 - SpiceKernels now supports displaying the contents of the headers of SPk/PCK files.
+- SpiceKernels can now return the available loaded SPK segments of an object.
 
 ### Changed